requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
albumentations==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
asttokens==2.0.5
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
boltons==23.0.0
brotlipy==0.7.0
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cmake==3.29.2
coloredlogs==15.0.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.0
cryptography==39.0.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
dpcpp-cpp-rt==2024.1.0
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.8.3
expecttest==0.1.4
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.9.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.47.2
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
gmpy2==2.1.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grad-cam==1.5.0
gradio==4.36.1
gradio_client==1.0.1
gradio_image_prompter==0.1.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hypothesis==6.75.2
idna==3.4
imageio==2.33.1
importlib_resources==6.4.0
imutils==0.5.4
intel-cmplr-lib-rt==2024.1.0
intel-cmplr-lic-rt==2024.1.0
intel-opencl-rt==2024.1.0
intel-openmp==2024.1.0
ipykernel==6.29.3
ipython==8.12.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
libarchive-c==2.9
librosa==0.10.1
lit==18.1.4
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2024.1.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml_collections==0.1.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.18.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.3
orjson==3.10.3
packaging==23.0
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==4.1.0
pluggy==1.0.0
pooch==1.8.0
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.12
pyOpenSSL==23.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-etcd==0.4.5
python-multipart==0.0.9
pytz==2022.7
PyYAML==6.0
pyzmq==25.1.2
qudida==0.0.4
referencing==0.35.1
rembg==2.0.57
requests==2.32.1
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruff==0.4.4
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
-e git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.7
stack-data==0.2.0
starlette==0.37.2
streamlit==1.36.0
streamlit-drawable-canvas==0.9.3
svgwrite==1.4.3
sympy==1.12
tbb==2021.12.0
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==1.0.8
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchelastic==0.2.2
torchtext==0.15.2
torchvision==0.15.2
tornado==6.1
tqdm==4.65.0
traitlets==5.7.1
triton==2.0.0
ttach==0.0.3
typer==0.12.3
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2023.4
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
zstandard==0.19.0